Enhance data model and database configuration
- Added `TrainerSpecializations` to `Account` and removed `CreatedTrainingSessions`.
- Updated `TargetWeight` types to `decimal?` in multiple classes.
- Introduced `TrainingSessionGoals` in `TrainingSession` and `Description` in `TrainingSessionGoal`.
- Updated project files to include Entity Framework Core packages.
- Created `DependencyInjection` class for SQL Server context configuration.
- Established `FitnessTrackerSqlServerDbContext` for managing entities.
- Added configuration classes for various entities to define schema.
- Created initial migration to set up the database schema.
- Updated model snapshot to reflect current database state.

+++ src/FitnessTracker.Domain/Entities/TrainingPlanGoal.cs
@@ -11,7 +11,7 @@ namespace FitnessTracker.Domain.Entities
     {
         public Guid Id { get; set; }
         public string Description { get; set; } = string.Empty;
-        public float? TargetWeight { get; set; }
+        public decimal? TargetWeight { get; set; }
         public EnumTrainingPlanGoalStatus Status { get; set; }
         public DateTime CreatedAt { get; set; }
