+++ src/FitnessTracker.Domain/Entities/WeightLog.cs
@@ -9,7 +9,7 @@ namespace FitnessTracker.Domain.Entities
     public class WeightLog
     {
         public Guid Id { get; set; }
-        public float Weight { get; set; }
+        public decimal Weight { get; set; }
         public string Unit { get; set; } = string.Empty;
         public DateTime CreatedAt { get; set; }
